Flink sql Problems
Query 1: Average idle time per hour per area
 

We need to find the average idle time, in minutes, taxis usually spend within a time window of one hour within a specific area. The final output for this query should be (areaId, avg(idleTime)) of the area within 1 hour.

Query 2: Frequent Routes
 

The goal of the query is to find the top 10 most frequent routes during the last 30 minutes. A route is represented by a starting areaID and an ending areaID. All routes completed within the last 30 minutes are considered for the query. A route is completed when we observe a ride start followed by a ride end for the same taxi.

Query 3: Profitable areas
 

The goal of this query is to identify the profitability per area for taxi drivers. The profitability of an area is determined by dividing the area profit by the number of empty taxis in that area within the last 15 minutes. The profit that originates from an area is computed by calculating the average fare + tip for trips that started in the area and ended within the last 15 minutes. The number of empty taxis in an area is the sum of taxis that had a drop-off location in that area less than 30 minutes ago and had no following pickup yet.